+++ CHANGELOG.md
@@ -52,6 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   leaks.
 - Correctly close input streams when ending a call while using a video
   transform device.
+- Resolve stream stop under low bitrate due to bps to kbps conversion.
 
 ## [2.6.2] - 2021-03-24